Grayskull harvesting support (#167)
In tt_grayskull.py there is an assert that says that Grayskull does not
support harvesting. However this is not the case, there are grayskull
cards with harvesting.
```python
def _handle_harvesting_for_nocTr_noc0_map(self, num_harvested_rows):
    assert num_harvested_rows == 0, "Harvesting not supported for Grayskull"
```

This code should be changed to this. This is the same solution found in
tt_blackhole.py file and it works for grayskull too.

```python
def _handle_harvesting_for_nocTr_noc0_map(self, num_harvested_rows):
    self.nocTr_x_to_noc0_x = {i: i for i in range(0, self.row_count())}
```