Exposing JTAG devices separately (#191)
This change completely splits UMD and JTAG "devices", so users will see
chips only over UMD (PCIe) or JTAG as "Devices" when using tt-lens. That
means the only one of the interfaces would be initialized and exposed
depending on the presence of the "--jtag" flag added recently.

Previously, chips that could also be accessed via JTAG, were accessed
using chip_id of UMD devices when using JTAG commands, which was not
ideal.