Fixing JTAG memdump functionality
Because it is slow and unnecessary to do dumping over JTAG, we decided
to remove all additional functionalities in future from JTAG (memdump,
sigdump) as they are either accessible via PCI or they are meaningless.

Base JTAG functionality will be preserved (NOC read/write + AXI
read/write) and will be used as fallback.