Added callstack (backtrace) support for metal kernels (#313)
Issue: https://github.com/tenstorrent/tt-exalens/issues/292

Now command takes as arguments list of elf paths and offsets. Offsets
represent location in memory where kernel is loaded (for firmware it
should be 0 or can be left empty), for kernels it should be
kernel_address variable from firmware code.

This change also relies on pyelftools patch regarding relocation support
for RISC-V. This patch was added to PyPl so now it is installed instead
of the original pyelftools until our changes are merged there.

Currently, it is not possible to load kernel directly from lens so we
rely on metal for that. Intended use is you run metal code, it hangs you
go to lens and use callstack command providing list of elf_paths for
kernels that are used and their offsets (kernel addresses).