Added read/write riscv memory to lib. (#323)
Issue: https://github.com/tenstorrent/tt-exalens/issues/241

Functions for reading and writing to riscv private memory are added to
library. To make use easier both of them use ensure_halted to halt program if not already halted and to continue it afterwards if it wasn't initially halted.

Added tests to test_lib.py