Simplify type hints (#394)
Closes #361 

Since we stopped supporting python3.8 we can utilize new features for
type hints such as list, dict, set.. instead of equivalents from typing
module. Also we can use | instead of Union and | None instead of
Optional. This PR deals with this.

*Note* In ttexalens/__init__.py switching from Union to | creates an
error so this is the only place where Union is left.

Some things (such as Callable, Iterable, TYPE_CHECKING...) does not have
identical alternatives so they are not changed.