Including risc debug tests for ETH cores (#403)
Closes #406 

Since chip no longer has to be reset after running risc debug tests on
ETH cores we can now include them.

Following tests fail on ETH cores, so there are skipped on them for now:

- test_invalidate_cache_with_nops_and_long_jump
- test_bne_with_debug_fail
- test_bne_without_debug
- test_bne_withou_debug_with_bp

Notably, last 3 tests set branch prediction using configuration register
which seems to not exist on ETH cores.

Also, reading PC through debug bus is not working on ETH cores (we
should add proper debug bus description for this) so it is disabled.

On blackhole, reading PC through gpr is not correct so checks for that
are disabled on blackhole.
We also need to read program base address from riscv private memory on
blackhole.