Enabling reading 32 bit formats from dest on blackhole (#533)
Closes #517
Also solves `blackhole` part of #467 and #518

Implemented reading dest directly from memory (`0xFFBD8000`) for 32 bit
formats (`int32` and `float32`). In this case to get correct values
reordering is not necessary (`int32` is printed as-is, `float32` is just
unpacked).

Also added printing for `uint16` for both `wormhole` and `blackhole`.
For `blackhole` also added printing for `int8`, `uint8` and `uint32`.

Added option for number of tiles to print in `dump_regfile` command. We
also now print `TILE ID` between tiles.

---------

Co-authored-by: Filip Vranic <fvranic@tenstorrent.com>