Using nanobind as base for ifc interface (#583)
Previously we had interface defined in `tt_exalens_ifc_base.py` and all
implementations had to override all methods.
With this change, we are exposing nanobind class as base interface and
with nanobind stub generation, we get verification with mypy as we had
before (it also works with pylance - VSCode editor default).
There is no need to verify bindings in a way we did earlier, because we
don't have wrapper class that forwards arguments back an forth.