Ignoring decode exception when logging GDB server response (#576)
Closes #585

While testing the GDB server, we encountered an issue where the server
failed to decode the response. The resulting exception was mistakenly
caught by a generic except block that was not intended for that part of
the code. To avoid this unintended behavior, the decoding exception is
now explicitly ignored.