Making _firmware_version in Device and telemetry_tags and telemetry_tag_ids cached_property to mitigate server bug (#626)
When active ETH is hung we are unable to obtain firmware version through
UMD (see issue: #620). This causes UMD to be stuck in infinite loop
essentially breaking our server functionality. This PR mitigates this by
only calling `get_firmware_version` when it is necessary (when we try to
read arc telemetry). To enable this `_firmware_version` in `Device`
class and `telemetry_tags` and `telemetry_tag_ids` in `ArcBlock` class
are now `cached_property` instead of fields.

`check_arc` script still will not work when active ETH is hung but other
scripts will.