Moving to marshal serializer in pyro5 (#627)
Problem was that `bytes` were serialized as `dict`. When receiving from
server, we could change result on client because it was result value.
When sending to server, we cannot change all arguments on server side,
so we switched to `marshal`.
This also solves `convert_bytes` from server as not needed...