Fix parse_string function in debug-bus command (#695)
The problem was that the reading of debug bus signals on Blackhole was
broken due to incorrect parsing when the `parse_string` function was
changed in the debug bus command. This has been fixed.

Also, the filenames where the commands are defined have been changed,
instead of containing ‘-’, they now use ‘_’, so that these files can be
imported. The command names remained the same by defining the long-name
attribute.