Expose reading N bytes from risc_debug (#773)
This PR adds two methods `read_memory_bytes` and `write_memory_bytes` to
`baby_risc_debug` and efectively moves the additional reading logic out
of `MemoryAccess` down, while solving the issue of `MemoryAccess` not
reading data from ncrisc as it only relied on debug hardware.

These two methods make alignments and fixups if the address is not
aligned so the caller does need to worry about that which in turn fixes
some stuff left in code from before (like when we overwrite words when
we write unaligned on Blackhole).

Closes #745