[elf variable] Proper error propagation on memory access violation (#789)
Currently, when running different operators on an elf variable if memory
it's pointing to is outside of L1/private memory for baby riscs (where
`RiscDebugMemoryAccess` is working in restricted mode) it will throw an
exception when reading its data. That exception is caught in different
operators and passed as a NotImplemented, which then confuses python (if
it's a > operator, it then runs the reverse which also throws the same
exception) and in the end we get an exception that doesn't really show
anyone that a memory access violation happened.

Now, for different operators I did the following:
1. For binary operators (comparison and arithmetic) - changed the
exception caught to a TypeError which is the one thrown from the inner
operator checked if the inner operator is not implemented (example for
LT we call read.value() < other and if those two types are incopatible
we'll get a TypeError), the Memory access error will pass through
2. For unary operators, NotImplemented doesn't really do anything as we
don't have reverse operations for python to use so I removed the
try_catch
3. Equal first checks array types then fallbacks to regular equal which
might through the memory access
4. Other operators are left as is

Added 2 test, first one validates all operators with memory access
violation with all permutations, second one validates type errors with
all permutations.

Closes #782