Extending dump-tensix-reg with rwc and adc making dump-tensix-state (#793)
Extending `dump-tensix-reg` command with RWCs and ADCs read from debug
bus and renaming it to `dump-tensix-state`.
Those groups are printed by default. If L1 address is provided we use
atomic group reading, if not we use newly added unsafe group reading
which reads group through debug hardware with 4 reads (4x32 bit =128
bit).
Similar to `TensixRegisterDescription` added `TensixDebugBusDescription`
that currently stores group names for RWC and ADC.

RWCs are currently not working on `blackhole` so we skip them on that
arch. #729