Promote MemoryAccess to a first-class concept (#808)
Currently, `MemoryAccess` is a concept defined int elf's `variable.py`
and used throughout the repo (and in some places in triage). The idea is
for it to evolve to be **the** abstraction for reading and writing data
to and from the memory and make those reads explicit.

The first step towards that is to promote it into a first-class concept
that will be exposed outside and inside exalens. Then it can be extended
with with different needs (like maybe a specialization called
`NocMemoryAccess` that reads only from NoC always, or from PCI/arc).

The next step is going through potential places for applying this change
and analyzing its requirements (like moving `register_store`'s)